Rollback to module version 1.3.1
There is one critical problem with versions 2.0+ - when the module
creates an apex A record e.g. infrahouse.com, it finds IP addresses of
network inyterfaces associated with the load balancer. Then it creates
the A record like:
```
$ host infrahouse.com
infrahouse.com has address 54.215.211.183
infrahouse.com has address 54.241.134.203
```
SO, the problem is the network interfaces can change.